+++ CustomEnchant/src/fr/enchantments/custom/implementation/Direct_Starvation.java
@@ -11,10 +11,10 @@ import fr.enchantments.custom.model.IDirectHitEnchantment;
 public class Direct_Starvation extends BaseEnchantment implements IDirectHitEnchantment
 {
 
-    public Direct_Starvation(String enchantmentName, int enchantmentID, int maxLevel) { super(enchantmentName, (short) enchantmentID, (short) maxLevel); }
+    public Direct_Starvation(String enchantmentName, int enchantmentID, int maxLevel) { super(enchantmentName, (short)enchantmentID, (short)maxLevel); }
 
 	@Override
-	public void onEntityHit(LivingEntity entityInflicter,LivingEntity entityVictim, ItemStack weaponUsed, short level, int damage) {
+	public void onEntityHit(LivingEntity entityInflicter,LivingEntity entityVictim, ItemStack weaponUsed, short level, short damage) {
 		entityVictim.addPotionEffect(new PotionEffect(PotionEffectType.HUNGER,level*100,level/3));
 	}
 }
